+++ src/app/src/components/DrawMap/MapFeatures.js
@@ -1,11 +1,13 @@
 import useAddPolygonCursor from './useAddPolygonCursor';
 import useEditingPolygon from './useEditingPolygon';
 import useGeocoderResult from './useGeocoderResult';
+import useTrackMapZoom from './useTrackMapZoom';
 
 export default function MapFeatures() {
     useEditingPolygon();
     useAddPolygonCursor();
     useGeocoderResult();
+    useTrackMapZoom();
 
     return null;
 }